Fixed #11 Querying phone numbers

+++ photos_api/tests.py
@@ -1,7 +1,6 @@
 import filecmp
 import httplib
 import json
-from django.contrib.auth import get_user_model
 from boto.s3.connection import S3Connection
 from boto.s3.key import Key
 from django.core.urlresolvers import reverse
@@ -10,15 +9,23 @@ import shutil
 
 from django.conf import settings
 from django.contrib import auth
-from django.test import TestCase
+from django.test import TestCase, Client
 from django.test.utils import override_settings
 from django.utils import timezone
 
-from phone_auth.models import PhoneNumber
+from phone_auth.models import PhoneNumber, PhoneContact, AnonymousPhoneNumber
 from photos.models import PendingPhoto, AlbumMember
+from photos_api import is_phone_number_mobile
 
 from photos_api.serializers import AlbumUpdateSerializer, MemberIdentifier, MemberIdentifierSerializer, AlbumAddSerializer
 import requests
+
+try:
+    from django.contrib.auth import get_user_model
+except ImportError:
+    from django.contrib.auth.models import User
+else:
+    User = get_user_model()
 
 
 @override_settings(LOCAL_PHOTO_BUCKETS_BASE_PATH='.tmp_photo_buckets')
@@ -68,6 +75,26 @@ class UserTest(BaseTestCase):
         self.assertFalse(auth.get_user_model().objects.filter(id=2).exists())
 
     def test_avatar_upload(self):
+        # ===================================
+        # Create data needed to test AnonymousPhoneNumber.avatar_file sync
+        amanda = User.objects.get(nickname='amanda')
+        fred = User.objects.get(nickname='fred')
+        apn = AnonymousPhoneNumber.objects.create(
+            phone_number='+12127189999',
+            is_mobile=True,
+            is_mobile_queried=timezone.now()
+        )
+        phone_contact = PhoneContact.objects.create(
+            anonymous_phone_number=apn,
+            user=amanda,
+            created_by_user=fred,
+            date_created=timezone.now(),
+            contact_nickname='amanda1'
+        )
+        initial_avatar=apn.avatar_file
+        # ===================================
+
+
         test_avatar_path = 'photos/test_photos/death-valley-sand-dunes.jpg'
         url = reverse('user-avatar', kwargs={'pk': 2})
 
@@ -82,15 +109,21 @@ class UserTest(BaseTestCase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response.headers['content-type'], "image/jpeg")
 
+        # Test that avatar for AnonymousPhoneNumber was changed as well
+        new_apn = AnonymousPhoneNumber.objects.get(pk=apn.pk)
+        self.assertEqual(user.avatar_file,
+                         new_apn.avatar_file)
+
         storage, bucket_name, filename = user.avatar_file.split(":")
 
         # Delete test image from S3
-        conn = S3Connection(settings.AWS_ACCESS_KEY,
-                            settings.AWS_SECRET_ACCESS_KEY)
-        bucket = conn.get_bucket(bucket_name)
-        key = Key(bucket, filename)
-        key.delete()
-        key.close(fast=True)
+        if storage == 's3':
+            conn = S3Connection(settings.AWS_ACCESS_KEY,
+                                settings.AWS_SECRET_ACCESS_KEY)
+            bucket = conn.get_bucket(bucket_name)
+            key = Key(bucket, filename)
+            key.delete()
+            key.close(fast=True)
 
     def test_user_updates(self):
         logged_in_user_id = 2
@@ -122,6 +155,112 @@ class UserTest(BaseTestCase):
             self.assertNotEqual(getattr(updated_user, attr_name),
                                 data[attr_name])
 
+    def test_query_phone_numbers_auth(self):
+        client = Client()
+        response = client.post(
+            reverse('query-phone-numbers'),
+            {'default_acountry': 'US', 'phone_numbers': []}
+        )
+        self.assertEqual(response.status_code, httplib.UNAUTHORIZED)
+
+    def test_query_phone_numbers_invalid(self):
+        data = {
+            "default_country": "us",
+            "phone_numbers": [
+                {
+                    "phone_number": "212718999923123123",
+                    "contact_nickname": "John Smith"
+                },
+            ]
+        }
+        response = self.client.post(reverse('query-phone-numbers'),
+                                    json.dumps(data, indent=4),
+                                    content_type="application/json")
+        self.assertEqual(response.status_code, httplib.OK)
+        response_data = json.loads(response.content)
+        self.assertEqual(response_data['phone_number_details'][0],
+                         {'phone_type': 'invalid'})
+
+    def test_query_phone_numbers_unknown(self):
+
+        # make sure we don't have this phone number
+        self.assertRaises(PhoneNumber.DoesNotExist, PhoneNumber.objects.get,
+                          phone_number='+12127189999')
+
+        data = {
+            "default_country": "us",
+            "phone_numbers": [
+                {
+                    "phone_number": "2127189999",
+                    "contact_nickname": "John Smith"
+                },
+            ]
+        }
+        response = self.client.post(reverse('query-phone-numbers'),
+                                    json.dumps(data, indent=4),
+                                    content_type="application/json")
+        self.assertEqual(response.status_code, httplib.OK)
+        response_data = json.loads(response.content)
+        phone_numbers = response_data['phone_number_details']
+        self.assertEqual(phone_numbers[0]['user_id'], None)
+        initial_avatar = phone_numbers[0]['avatar_url']
+
+
+        # Make request from different user and with the same anonymous phone
+        # number and make sure the same avatar is used.
+        client = Client()
+        client.login(username=3, password="barney")
+        response = client.post(reverse('query-phone-numbers'),
+                               json.dumps(data, indent=4),
+                               content_type='application/json')
+        self.assertEqual(response.status_code, httplib.OK)
+        response_data = json.loads(response.content)
+        phone_numbers = response_data['phone_number_details']
+        self.assertEqual(phone_numbers[0]['user_id'], None)
+        self.assertEqual(phone_numbers[0]['avatar_url'], initial_avatar)
+
+    def test_query_phone_numbers_existing(self):
+
+        amanda = User.objects.get(nickname='amanda')
+        fred = User.objects.get(nickname='fred')
+        phone_number = PhoneNumber.objects.create(
+            phone_number='+12127189999',
+            user=fred,
+            date_created=timezone.now(),
+            verified=True
+        )
+
+        data = {
+            "default_country": "us",
+            "phone_numbers": [
+                {
+                    "phone_number": "2127189999",
+                    "contact_nickname": "John Smith"
+                },
+            ]
+        }
+        response = self.client.post(reverse('query-phone-numbers'),
+                                    json.dumps(data, indent=4),
+                                    content_type="application/json")
+        self.assertEqual(response.status_code, httplib.OK)
+        response_data = json.loads(response.content)
+        phone_numbers = response_data['phone_number_details']
+        self.assertEqual(phone_numbers[0]['user_id'], fred.id)
+        initial_avatar = phone_numbers[0]['avatar_url']
+
+
+        # Make request from different user and with the same anonymous phone
+        # number and make sure the same avatar is used.
+        client = Client()
+        client.login(username=3, password="barney")
+        response = client.post(reverse('query-phone-numbers'),
+                               json.dumps(data, indent=4),
+                               content_type='application/json')
+        self.assertEqual(response.status_code, httplib.OK)
+        response_data = json.loads(response.content)
+        phone_numbers = response_data['phone_number_details']
+        self.assertEqual(phone_numbers[0]['user_id'], fred.id)
+        self.assertEqual(phone_numbers[0]['avatar_url'], initial_avatar)
 
 class NotModifiedTest(BaseTestCase):
     def setUp(self):
@@ -570,3 +709,15 @@ class MembersTests(BaseTestCase):
         # Retrieving album details is forbidden, because user is not a member of the album anymore
         album_after_response = self.client.get(reverse('album-detail', kwargs={'pk':9}))
         self.assertEqual(album_after_response.status_code, httplib.FORBIDDEN)
+
+
+class TestPhoneNumberMobile(TestCase):
+    def test_1(self):
+        # "97254", "97252", "97250", "97258" These are prefixes for mobile
+        # numbers in Israel
+        self.assertTrue(is_phone_number_mobile("+972541231212", 'IL'))
+        self.assertTrue(is_phone_number_mobile("+972521231212", 'IL'))
+        self.assertTrue(is_phone_number_mobile("+972501231212", 'IL'))
+        self.assertTrue(is_phone_number_mobile("+972581231212", 'IL'))
+        self.assertFalse(is_phone_number_mobile("+972231231212", 'IL'))
+        self.assertTrue(is_phone_number_mobile("+12127184000", 'US'))
